requirements.txt
flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
gunicorn
psycopg2==2.8.5
python-dotenv==0.14.0
pytest-flask
pytest-mock